requirements.txt
beautifulsoup4==4.13.4
ctklistbox==1.5
customtkinter==5.2.1
mss==10.0.0
numpy==2.3.1
opencv_python==4.11.0.86
Pillow==11.3.0
psutil==5.9.5
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pymem==1.13.0
pynput==1.7.7
pyperclip==1.8.2
pytesseract==0.3.13
pytest==8.3.4
pywin32==306
